Fix Windows release cross-build
Build libsodium for MinGW into a project-local .deps prefix instead of /usr so GitHub runners do not need system write access.

Route the Windows client through the shared sodium CMake target, expose sodium headers through epn-crypto, add missing release workflow tools, and remove a SOCKS5 narrowing warning.

+++ apps/epn-win-client/main.cpp
@@ -401,7 +401,7 @@ socks5_handshake(std::shared_ptr<tcp::socket> sock) {
 }
 
 static void socks5_reply(std::shared_ptr<tcp::socket> sock, bool ok) {
-    uint8_t rep[10]={0x05, ok?0x00u:0x05u, 0x00, 0x01,0,0,0,0,0,0};
+    uint8_t rep[10]={0x05, static_cast<uint8_t>(ok ? 0x00u : 0x05u), 0x00, 0x01,0,0,0,0,0,0};
     std::error_code ec; asio::write(*sock,asio::buffer(rep),ec);
 }
